Adds class for 2D texture views
